history state not working for scxml (AISingleChoiceElement) All tests running for .rb version of the statemachines

+++ lib/MINT-core/model/aui/AIChoice.rb
@@ -2,10 +2,9 @@ module MINT
   class AIChoice < AIC
     def initialize_statemachine
       if @statemachine.blank?
-=begin
         parser = StatemachineParser.new(self)
         @statemachine = parser.build_from_scxml "lib/MINT-core/model/aui/aichoice.scxml"
-=end
+=begin
  @statemachine = Statemachine.build do
           trans :initialized,:organize, :organized
           trans :organized, :present, :p_t
@@ -39,23 +38,26 @@ module MINT
               end
               statemachine :s2 do
                 superstate :l do
-                  trans :listing, :drop, :listing, :add_element, Proc.new{In(:focus)}
+                  trans :listing, :drop, :listing, :add_element
                 end
               end
             end
           end
         end
+=end
       @statemachine.reset
       end
     end
     def add_element
-      f = AISingleChoiceElement.first(:new_states=> /#{Regexp.quote("dragging")}/)
-      self.childs << f
-      self.childs.save
-      f.process_event(:drop)
-      f.process_event(:focus)
-      self.process_event(:defocus)
-      f.destroy
+      if self.is_in? :focused
+        f = AISingleChoiceElement.first(:new_states=> /#{Regexp.quote("dragging")}/)
+        self.childs << f
+        self.childs.save
+        f.process_event(:drop)
+        f.process_event(:focus)
+        self.process_event(:defocus)
+        f.destroy
+      end
     end
   end
